change handling of group actions of nested groups: children groups inherit actions from parent groups instead of overwriting them, if parent and child use the same action both have an effect
add flag to menu-items indicating whether they refer to an action of the top group or not, those items are hidden by the client for floating tasks now, because they are shown in the group's toolbar anyhow

For instance in my text editor, the group's menu shows global actions (new, open, quit, ...) while a floating editor's menu only shows actions related to a single file (save, close, ...). If the editor is pinned and focused the menu includes all actions. This solves the problem of the previous behaviour with which global actions disappeared in that situation.

git-svn-id: https://svn.cs.ru.nl/repos/iTask-system/trunk@1081 63da3aa8-80fd-4f01-9db8-e6ea747a3da2