Add compute combinator
Compute is a return that returns a value that possibly takes long to
calculate, thus blocking the execution. The compute version of return
will display a loader with the given text to at least notify the user of
blocking computations.
Right now it's implemented hacky by waiting one second to make sure the
loader UI is sent to the client.