Change sds rewriting (WIP: segfaults)
Change the way shares are rewritten. Each share now rewrites itself
instead of its children, and returns the rewritten variant when an
operation yields a result. This ensures in the case of a modify that we
can first read and then use the returned share to execute the write.

Why?
- Giving the modify function an extra parameter with the complete share,
which can be used after reading is done, is no longer sufficient due to
the implementation of the SDSLens/SDSParallel.
- Implementing a recover class which can transform a after-read-sds to
the original is not possible due to typing constraints.

This approach specifically allows the SDSRemoteSource/SDSRemoteService
to rewrite itself from SDSQueued to SDSValue, while moving along the
underlying SDS definition.